added skeleton for all other robots
