+++ poutine/pubsub/pubsub.go
@@ -4,3 +4,5 @@ type PubSub interface {
 	Publish(channel, msg string) error
 	Subscribe(channel string, mh MessageHandler) error
 }
+
+type MessageHandler func(string) error
